requirements.txt
redis==3.5.3
redlock-py==1.0.8
PyYAML==5.4.1